Consolidate db utils into index
