--- src/database/overwrite-db.ts
@@ -1,5 +1,0 @@
-import { useDbFile } from "@/database/client";
-
-export async function overwriteDb() {
-  const { overwriteDatabaseFile } = useDbFile();
-}
