+++ src/routes/test/+page.ts
@@ -1,4 +1,4 @@
-import { exposeDb } from "@/database/expose-db";
+import { exposeDb } from "@/database";
 
 export const ssr = false;
